refactor: master 브랜치 충돌 병합

+++ src/main/java/com/onedrinktoday/backend/global/cache/CacheService.java
@@ -12,7 +12,7 @@ public class CacheService {
   private final PostRepository postRepository;
 
   @Cacheable(key = "#drinkId", value = "avg-rating")
-  public Float getAverageRating(Long drinkId) {
+  public Double getAverageRating(Long drinkId) {
     return postRepository.getAverageRating(drinkId);
   }
 }
